move components into expected directory
