--- src/assets/components/DrawButton.jsx
@@ -1,7 +1,0 @@
-// import React from 'react'; => am using react version 18.3.1 so don't neeed to import react 
-
-const DrawButton = ({ drawNames }) => {
-    return <button onClick={drawNames}>Draw Names🎅🧑‍🎄</button>;
-};
-
-export default DrawButton;
