--- src/assets/components/ParticipantList.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const ParticipantList = ({ participants }) => {
-    return (
-        <ul>
-            {participants.map((participant) => (
-                <li key={participant}>{participant}</li>
-            ))}
-        </ul>
-    );
-};
-
-export default ParticipantList;
