+++ src/components/DrawButton.jsx
@@ -1,0 +1,5 @@
+const DrawButton = ({ drawNames }) => {
+  return <button onClick={drawNames}>Draw Names🎅🧑‍🎄</button>;
+};
+
+export default DrawButton;
