+++ src/components/ParticipantList.jsx
@@ -1,0 +1,11 @@
+const ParticipantList = ({ participants }) => {
+  return (
+    <ul>
+      {participants.map((participant) => (
+        <li key={participant}>{participant}</li>
+      ))}
+    </ul>
+  );
+};
+
+export default ParticipantList;
