+++ src/components/ResultList.jsx
@@ -1,0 +1,13 @@
+const ResultList = ({ results = [] }) => {
+  return (
+    <ul>
+      {results.map((result, index) => (
+        <li key={index}>
+          {result.giver} {"→"} {result.receiver}
+        </li>
+      ))}
+    </ul>
+  );
+};
+
+export default ResultList;
